+++ R/globals.R
@@ -1,3 +1,4 @@
 # https://community.rstudio.com/t/how-to-solve-no-visible-binding-for-global-variable-note/28887
 # https://www.r-bloggers.com/2019/08/no-visible-binding-for-global-variable/
-utils::globalVariables(c("prevalence_1", "prevalence_2", "prevalence_pct", "FP", "NPV", "PPV", "Prevalence", "FN", "value", "prevalence", "sensitivity", "specificity"))
+utils::globalVariables(c("prevalence_1", "prevalence_2", "prevalence_pct", "FP", "NPV", "PPV", "Prevalence", "FN", "value", "prevalence", "sensitivity", "specificity",
+                         "N", "classification", "fill_str", "test_result", "type"))